fix: types dist/tokens
