+++ packages/types/src/index.ts
@@ -18,5 +18,5 @@ export type { Runtime } from './runtime'
 export type { AnyFunction, Artifact, Dict, RequiredBy } from './shared'
 export type { StaticCssOptions } from './static-css'
 export type { CssKeyframes, GlobalStyleObject, SystemStyleObject } from './system-types'
-export type { SemanticToken, SemanticTokens, Token, TokenCategory, TokenDataTypes, Tokens } from '../tokens'
+export type { SemanticToken, SemanticTokens, Token, TokenCategory, TokenDataTypes, Tokens } from './tokens'
 export type { PropertyConfig, PropertyTransform, PropertyValues, UtilityConfig } from './utility'
